pyproject.toml
[project]
name = "astra-core"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.24.0",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pyswisseph>=2.10.3.2",
    "python-dotenv>=1.2.2",
    "pytz>=2025.2",
    "sqlalchemy>=2.0.48",
    "timezonefinder>=8.2.1",
    "uvicorn>=0.41.0",
]
